jooby apt: generate source code ref #2968
- more code cleanup
