+++ modules/jooby-apt/src/main/resources/META-INF/gradle/incremental.annotation.processors
@@ -1,1 +1,1 @@
-io.jooby.apt.MvcSourceCodeProcessor,dynamic
+io.jooby.apt.JoobyProcessor,dynamic
