+++ modules/jooby-apt/src/main/resources/META-INF/services/javax.annotation.processing.Processor
@@ -1,1 +1,1 @@
-io.jooby.apt.MvcSourceCodeProcessor
+io.jooby.apt.JoobyProcessor
