+++ modules/jooby-apt/src/test/java/tests/Issue1545.java
@@ -10,14 +10,14 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import org.junit.jupiter.api.Test;
 
 import io.jooby.StatusCode;
-import io.jooby.apt.NewProcessorRunner;
+import io.jooby.apt.ProcessorRunner;
 import io.jooby.test.MockRouter;
 import source.Controller1545;
 
 public class Issue1545 {
   @Test
   public void shouldSetNoContentCodeForVoidRoute() throws Exception {
-    new NewProcessorRunner(new Controller1545())
+    new ProcessorRunner(new Controller1545())
         .withRouter(
             app -> {
               MockRouter router = new MockRouter(app);
