+++ modules/jooby-apt/src/test/java/tests/Issue1552.java
@@ -13,7 +13,7 @@ import java.util.Arrays;
 import org.junit.jupiter.api.Test;
 
 import io.jooby.Context;
-import io.jooby.apt.NewProcessorRunner;
+import io.jooby.apt.ProcessorRunner;
 import io.jooby.test.MockRouter;
 import source.Controller1552;
 import source.Controller1552Empty;
@@ -22,7 +22,7 @@ import source.JavaBeanParam;
 public class Issue1552 {
   @Test
   public void inherited() throws Exception {
-    new NewProcessorRunner(new Controller1552())
+    new ProcessorRunner(new Controller1552())
         .withRouter(
             app -> {
               MockRouter router = new MockRouter(app);
@@ -55,7 +55,7 @@ public class Issue1552 {
 
   @Test
   public void inherited_empty() throws Exception {
-    new NewProcessorRunner(new Controller1552Empty())
+    new ProcessorRunner(new Controller1552Empty())
         .withRouter(
             app -> {
               MockRouter router = new MockRouter(app);
