+++ modules/jooby-apt/src/test/java/tests/i2405/Issue2405.java
@@ -9,14 +9,14 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import org.junit.jupiter.api.Test;
 
-import io.jooby.apt.NewProcessorRunner;
+import io.jooby.apt.ProcessorRunner;
 import io.jooby.test.MockContext;
 import io.jooby.test.MockRouter;
 
 public class Issue2405 {
   @Test
   public void shouldGenerateUniqueNames() throws Exception {
-    new NewProcessorRunner(new C2405())
+    new ProcessorRunner(new C2405())
         .withRouter(
             app -> {
               app.converter(new Converter2405());
