+++ modules/jooby-apt/src/test/java/tests/i2408/Issue2408.java
@@ -10,7 +10,7 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import org.junit.jupiter.api.Test;
 
-import io.jooby.apt.NewProcessorRunner;
+import io.jooby.apt.ProcessorRunner;
 import io.jooby.exception.MissingValueException;
 import io.jooby.test.MockContext;
 import io.jooby.test.MockRouter;
@@ -18,7 +18,7 @@ import io.jooby.test.MockRouter;
 public class Issue2408 {
   @Test
   public void shouldNotIgnoreAnnotationOnParam() throws Exception {
-    new NewProcessorRunner(new C2408())
+    new ProcessorRunner(new C2408())
         .withRouter(
             app -> {
               MockRouter router = new MockRouter(app);
