+++ modules/jooby-apt/src/test/java/tests/i2417/Issue2417.java
@@ -9,14 +9,14 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import org.junit.jupiter.api.Test;
 
-import io.jooby.apt.NewProcessorRunner;
+import io.jooby.apt.ProcessorRunner;
 import io.jooby.test.MockContext;
 import io.jooby.test.MockRouter;
 
 public class Issue2417 {
   @Test
   public void shouldNotIgnoreAnnotationOnParam() throws Exception {
-    new NewProcessorRunner(new C2417())
+    new ProcessorRunner(new C2417())
         .withRouter(
             app -> {
               MockRouter router = new MockRouter(app);
