+++ modules/jooby-apt/src/test/java/tests/i2525/Issue2525.java
@@ -10,14 +10,14 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import org.junit.jupiter.api.Test;
 
 import io.jooby.Context;
-import io.jooby.apt.NewProcessorRunner;
+import io.jooby.apt.ProcessorRunner;
 import io.jooby.test.MockContext;
 import io.jooby.test.MockRouter;
 
 public class Issue2525 {
   @Test
   public void shouldAcceptEmptyList() throws Exception {
-    new NewProcessorRunner(new C2525())
+    new ProcessorRunner(new C2525())
         .withRouter(
             app -> {
               var router = new MockRouter(app);
