+++ modules/jooby-apt/src/test/java/tests/i2629/Issue2629b.java
@@ -10,7 +10,7 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 
 import org.junit.jupiter.api.Test;
 
-import io.jooby.apt.NewProcessorRunner;
+import io.jooby.apt.ProcessorRunner;
 import io.jooby.test.MockContext;
 import io.jooby.test.MockRouter;
 
@@ -18,7 +18,7 @@ public class Issue2629b {
 
   @Test
   public void shouldSetMvcMethod() throws Exception {
-    new NewProcessorRunner(new C2629b())
+    new ProcessorRunner(new C2629b())
         .withRouter(
             app -> {
               MockRouter router = new MockRouter(app);
